Initial Commit
This is the start of a beaglebone black based 8 channel pwm fan
controller. Right now it's just the board design and schematic. While
everything is wired up at this point it likely will need a bunch of
refinement. Some of the net paths are less than ideal, it's not clear
how much that will mater though because PWM fans operate at only ~25kHz.
In the future software will be added to this repo as well.